// src/generated/prisma/enums.ts
export const UserRole = {
  SUPER_ADMIN: 'SUPER_ADMIN',
  ADMIN: 'ADMIN',
  OWNER: 'OWNER',
  EMPLOYEE: 'EMPLOYEE',
  SUPPLIER: 'SUPPLIER',
  CUSTOMER: 'CUSTOMER'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const UserStatus = {
  BLOCKED: 'BLOCKED',
  ACTIVE: 'ACTIVE',
  PENDING: 'PENDING',
  DELETED: 'DELETED'
} as const

export type UserStatus = (typeof UserStatus)[keyof typeof UserStatus]